Read all inputs instead of those for a particular device.
This means I don't have to keep changing the script when I plugin my
external keyboard. The downside is that test-xi2 seems to open a random
window (that can be closed). Closing a random window once on startup
seems like a small price to pay compared to having to keep editing the
script (and restarting it, which keeps creating zombie threads).